Hi Jose,
> No, I haven't touch them, so you will see that alerts because they are
> pointing to my sql install file. Did you see my previous message where i
> talk (again) of the sessionid?
> I don't want to touch them until we definitively close the database design.
Yes, you are right. I made quick changes, so that the window open now without failure. It slipped my mind because I did not built this models. But I'll definitely forget it again :)

> The media folder is included in the xml file (lines 32 to 37).
Sorry, my look in the morning yesterday was to quick or I was to blind to see the media entry ... everything is OK ...

> Ok. Basically I added the javascript function to load the pop-up when
> the document is ready into a file (onpageload.js); this why I call the
> file and the modal pop-up is loaded.

I do not understand why you have to click twice on the close button to close the modal window. Do you have an explanation for this behaviour?
What do you want to write exactly on the modal window?

I hope your daughter is well and healthy again.

Regards
Astrid