$doc = & JFactory::getDocument();
I change it to

$doc = JFactory::getDocument();

From PHP5 onwards objects are assigned by reference, so there is no need
to do that explicitly with an ampersand. Is this right?

I do not think that many people want to override our extension, but I am
learning and I read this:
https://www.babdev.com/blog/139-use-the-media-folder-allow-overridable-media
That's why I want to change things like

$doc->addScript($this->media_path . '/javascript/JoommarkSetTimeout.js');

to

JHtml::script('com_joommark/javascript/JoommarkSetTimeout.js', false, true);